fix: pass timeout arg to server
Refactors how we pass url params to the API, all args are now normalised
in one place.  If a timeout is passed it is passed on to the server as
a url search param.

N.b. this does mean that timeouts are set in the client and on the server
so it's not always clear which will expire first, though one will expire.

We do our best to ensure the errors are the same, though when the API
is streaming there's no way to tell if the request completed successfully
or if the timeout occured as the timeout message arrives in http trailers.